code/requirements.txt
fonttools==4.25.0
imageio==2.21.3
numpy==1.23.3
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.4.3
Pillow==9.2.0
scikit-image==0.19.3
scikit-learn 
scikit-plot==0.3.7
scipy==1.9.0
seaborn
Shapely==1.8.4
tensorboard
termcolor==1.1.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm